Separating tests into different files
